Move k8s client package into top level for inclusion in other projects
