--- internal/tools/k8s/deployments_test.go
@@ -1,72 +1,0 @@
-// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
-// See LICENSE.txt for license information.
-//
-
-package k8s
-
-import (
-	"testing"
-
-	"github.com/stretchr/testify/require"
-	appsv1 "k8s.io/api/apps/v1"
-	appsbetav1 "k8s.io/api/apps/v1beta1"
-	appsv1beta2 "k8s.io/api/apps/v1beta2"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-)
-
-func TestDeploymentsV1(t *testing.T) {
-	testClient := newTestKubeClient()
-	deployment := &appsv1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{Name: "test-deployment"},
-	}
-	namespace := "testing"
-
-	t.Run("create deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentV1(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-	t.Run("create duplicate deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentV1(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-}
-
-func TestDeploymentsBetaV1(t *testing.T) {
-	testClient := newTestKubeClient()
-	deployment := &appsbetav1.Deployment{
-		ObjectMeta: metav1.ObjectMeta{Name: "test-deployment"},
-	}
-	namespace := "testing"
-
-	t.Run("create deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentBetaV1(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-	t.Run("create duplicate deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentBetaV1(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-}
-
-func TestDeploymentsBetaV2(t *testing.T) {
-	testClient := newTestKubeClient()
-	deployment := &appsv1beta2.Deployment{
-		ObjectMeta: metav1.ObjectMeta{Name: "test-deployment"},
-	}
-	namespace := "testing"
-
-	t.Run("create deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentBetaV2(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-	t.Run("create duplicate deployment", func(t *testing.T) {
-		result, err := testClient.createOrUpdateDeploymentBetaV2(namespace, deployment)
-		require.NoError(t, err)
-		require.Equal(t, deployment.GetName(), result.GetName())
-	})
-}
